+++ game/Content/Locations/Ohio/Ohio.py
@@ -4,6 +4,7 @@ from Location.Location import Location
 
 # content packages
 from .OhioZooAndAquarium.OhioZooAndAquarium import OhioZooAndAquarium
+from .OhioCity.OhioCity import OhioCity
 
 
 class Ohio(Location):
@@ -12,5 +13,6 @@ class Ohio(Location):
             "Ohio",
             [
                 OhioZooAndAquarium(),
+                OhioCity()
             ]
         )
