+++ game/Content/Locations/Ohio/OhioCity/Weapons/BlueMeth.py
