+++ game/Content/Locations/Ohio/OhioCity/Weapons/Hands.py
