+++ game/Content/Locations/Ohio/OhioCity/Weapons/ImpostersKnife.py
