+++ game/Content/Locations/Ohio/OhioCity/Weapons/LiverBoostedHands.py
